Made activator perform a simple activation on a cube.
I made it so that the cube in the corner changes to green when the activator is near enough. I wasn't sure of a good way to abstract, so as of now the plan is to hard code in everything we can activate to do the specific thing it does when activated. Visual representation of the activator's range will come next, stay tuned.
