+++ Assets/Prefabs/Activatable Cube.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ddf813b4cfba1864180b1f18d8ba3a1c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
