+++ ProjectSettings/TagManager.asset
@@ -7,6 +7,7 @@ TagManager:
   - Liftable
   - HumanDim
   - OtherDim
+  - Activatable
   layers:
   - Default
   - TransparentFX
